% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_api.R
\name{get_functional_enrichment}
\alias{get_functional_enrichment}
\title{Get functional enrichment.}
\usage{
get_functional_enrichment(identifiers, species_ncbi_id, format = "tsv",
  caller_identity = "TenzerLab", ...)
}
\arguments{
\item{identifiers}{Character array of protein identifiers or stringIDs.}

\item{species_ncbi_id}{The NCBI identification of the organism.}

\item{format}{Change to 'json' or 'xml' if 'tsv' does not work.}

\item{...}{Other settings listed in}
}
\value{
data.table with translated entries.
}
\description{
Get functional enrichment.
}
