% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_api.R
\name{get_string_db_data_}
\alias{get_string_db_data_}
\title{General method to connect to STRING-DB and retrieve output.}
\usage{
get_string_db_data_(x, method, format = "tsv", ...)
}
\arguments{
\item{format}{Change to 'json' or 'xml' if 'tsv' does not work.}

\item{identifiers}{Character array of protein identifiers.}
}
\value{
POST object: the content field contains binary outputs.
}
\description{
General method to connect to STRING-DB and retrieve output.
}
