% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncbi_ids}
\alias{ncbi_ids}
\title{ncbi_ids}
\format{
A data frame
\describe{
  \item{entry}{Long uniprot id.}
  \item{uniprot}{Short uniprot id.}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
ncbi_ids
}
\description{
Magnaporthe Carysae, NCBI ID 318829-
}
\keyword{datasets}
